requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
billiard==4.2.1
blessed==1.20.0
cachetools==5.3.1
celery==5.4.0
certifi==2023.5.7
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.7
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
dj-database-url==2.3.0
Django==4.2.2
django-allauth==0.54.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-filter==25.1
django-livesync==0.5
django-oauth-toolkit==2.3.0
django-picklefield==3.2
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
filelock==3.12.2
google-auth==2.22.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
jsonpickle==0.9.6
jwcrypto==1.5.0
kombu==5.4.2
loguru==0.7.3
lxml==5.3.1
mailtrap==2.0.1
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
pillow==11.1.0
platformdirs==3.7.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.7.0
pypdf==5.3.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
qrcode==8.0
redis==3.5.3
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==75.8.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.4.0
tornado==6.3.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
uritools==4.0.3
urllib3==1.26.16
vine==5.1.0
virtualenv==20.23.1
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
whitenoise==6.5.0
wrapt==1.15.0
xhtml2pdf==0.2.17
